Ändrat från int till short i CRUD

+++ src/main/java/com/sandstrom/crudOperations/CrudOfAddress.java
@@ -120,8 +120,8 @@ public class CrudOfAddress implements CrudOfAddressInterface{
         return latestId != null ? latestId : 0;
     }
 
-    private int getGeneratedId(){
-        return idGenerating() + 1;
+    private short getGeneratedId(){
+        return (short) (idGenerating() + 1);
     }
